% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holor-package.R
\docType{package}
\name{holor-package}
\alias{holor}
\alias{holor-package}
\title{Render Quilt Images for the Looking Glass Holographic Display}
\description{
An open-source (GPL-3) R package to render quilt images for the holographic
display of the Looking Glass Factory.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DvP17/holor}
  \item Report bugs at \url{https://github.com/DvP17/holor/issues}
}

}
\keyword{internal}
