% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_quilt.R
\name{preview_quilt}
\alias{preview_quilt}
\title{Preview quilts}
\usage{
preview_quilt(file, width = 1200, fps = 20, save = FALSE)
}
\arguments{
\item{file}{File name of the quilt image.}

\item{width}{Width of the GIF. Default is 1200.}

\item{fps}{Frames per second. Default is 20.}

\item{save}{If GIF should be saved. Default is FALSE.}
}
\description{
\code{preview_quilt} creates a GIF preview of a quilt image.
}
