% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quilt.R
\name{quilt}
\alias{quilt}
\title{Render quilts}
\usage{
quilt(fun, cone = 10, theta = 0, phi = 0, device = "LGP", file)
}
\arguments{
\item{fun}{The plot function. Currently supported for \code{plot3D::scatter3D} and \code{plot3D::hist3D}.}

\item{cone}{The viewing cone.}

\item{theta}{The theta parameter.}

\item{phi}{The phi parameter.}

\item{device}{The Looking Glass Device.}
}
\description{
\code{quilt} renders quilt images for the Looking Glass holographic display.
}
