% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quilt.R
\name{quilt}
\alias{quilt}
\title{Create a quilt image}
\usage{
quilt(cone = 10, theta = 0, phi = 0, device = "LGP", file)
}
\arguments{
\item{cone}{The viewing cone}

\item{theta}{Is theta}

\item{phi}{Is phi}

\item{device}{Is Looking Glass Device}
}
\description{
Create a quilt image
}
